+++ src/dlls/engine/2/2.c
@@ -3404,7 +3404,7 @@ int objSeqExecCmd06(GameObject* obj, GameObject* sourceObj, u8* seq, int cmd, s8
         gameTimerStop();
         break;
     case 13:
-        Sfx_StopObjectChannel((u32)sourceObj, 0x7f);
+        Sfx_StopObjectChannel(sourceObj, 0x7f);
         break;
     case 16:
         *(s8*)&((ObjSeqState*)seq)->unk7D = cmdArg;
@@ -4737,11 +4737,11 @@ int ObjSeq_ExecuteActionCommand(GameObject* obj, u8* action, u8** cmdPtr, s8 fla
         }
         if (((*(s16*)(cmd + 2) >> 12) & 0xf) != 0xf)
         {
-            Sfx_PlayFromObject((u32)obj, (u16)(*(s16*)(cmd + 2) & 0xfff));
+            Sfx_PlayFromObject(obj, (u16)(*(s16*)(cmd + 2) & 0xfff));
         }
         else
         {
-            Sfx_PlayFromObject((u32)obj, (u16)(*(s16*)(cmd + 2) & 0xfff));
+            Sfx_PlayFromObject(obj, (u16)(*(s16*)(cmd + 2) & 0xfff));
             ((ObjSeqState*)seq)->sfxTimer[3] = -1;
             ((ObjSeqState*)seq)->sfxId[3] = (s16)(*(s16*)(cmd + 2) & 0xfff);
         }
@@ -5350,15 +5350,15 @@ void ObjSeq_ApplyFrameCurves(GameObject* obj, GameObject* seqObj, u8* seq, int f
         {
             if (*(s16*)(seq + i * 2 + 0x30) != 0)
             {
-                Sfx_IsPlayingFromObject((u32)seqObj, (u16) * (s16*)(seq + i * 2 + 0x38));
+                Sfx_IsPlayingFromObject(seqObj, (u16) * (s16*)(seq + i * 2 + 0x38));
             }
         }
 
         if (vol > 0 && ((ObjSeqState*)seq)->sfxTimer[3] != 0)
         {
-            if (Sfx_IsPlayingFromObject((u32)seqObj, (u16)((ObjSeqState*)seq)->sfxId[3]) != 0)
+            if (Sfx_IsPlayingFromObject(seqObj, (u16)((ObjSeqState*)seq)->sfxId[3]) != 0)
             {
-                Sfx_SetObjectSfxVolume((u32)seqObj, (u16)((ObjSeqState*)seq)->sfxId[3], vol, 0.5f);
+                Sfx_SetObjectSfxVolume(seqObj, (u16)((ObjSeqState*)seq)->sfxId[3], vol, 0.5f);
             }
         }
